add a bean validation context to control property validation (#317)
Sometimes you need to control what properties are validated and there is no current API to do that.

The use case if partial updates. So you want to send a request that partially updates an object in the database so you need to validate only a subset of the properties. Currently you can kind of do this with groups but it is awkward and requires static declaration.

The PR adds a BeanValidationContext which can be used to control property inclusions and perhaps for other purposes in the future. The code that passed around the groups has been refactored to so with the context instead.