Remove Introspected from MyValidatorBean example
The docs are updated to no longer show the use of the Introspected
annotation on the MyValidatorBean example, as it cannot successfully
be used on a validator that is a managed bean with dependencies.

DefaultInternalConstraintValidatorFactory is updated to include the
exception cause in the stack trace when a validator cannot be
initialized.

The CustomMessageConstraint2Validator test class for
CustomConstraintsSpec is updated to include a dependency that must
successfully be injected in order to verify that this works as long as
the validator is not also Introspected.