Ensure validators accessible as beans from usingContext
The usingContext method of DefaultValidatoryFactory is updated to check
for an existing non-null ValidatorConfiguration that is itself an
instance of ValidatorContext (which will always be the case when
created via DI).

This allows custom validation constraints to be located in the
BeanContext when validation is invoked by a Jakarta validation
implementation.